Create others_customizations.js
Moved the back/forward buttons to personal bar
Since I only use the sidebar-button for bookmarks side bar, I changed the icon with CSS somewhere else, and here I changed the text.
A change for the Noscript button because in "icon+text" mode, it was uselessly taking space, and the tooltip still provide the full info in a cleaner way anyway.
A fix for the webgame NationStates, because I have many nations, and didn't want to scroll. Attempted to bugfix some issues caused by that code, but maybe users with different sites would see issues appearing from it.